Added the procedures to database, modified the  of the user control.
Two procedures created, the first one get the ID of the productModel and the other one loads all the mains objects (id, name, price and image), and this one gets the id randomly from the other method.

Aesthetics changes to the user control: background color changed and added some borders.